Made some changes to login screen
Added a submit button for when user is ready to attempt to login, changed the input label to hide the characters with stars. If password matches "password" (Will change later to users password), changes screens to screen after login. If password is wrong, started a counter that will reset passwords after 10 attempts.